can perform comparison now

+++ RGBDFusion/include/ImageProc.h
@@ -19,7 +19,7 @@ void RenderScene(const cv::cuda::GpuMat vmap, const cv::cuda::GpuMat nmap, cv::c
 void ComputeImageGradientCentralDifference(const cv::cuda::GpuMat image, cv::cuda::GpuMat &gx, cv::cuda::GpuMat &gy);
 
 void TransformReferencePoint(const cv::cuda::GpuMat depth, cv::cuda::GpuMat &vmap,
-                             const Eigen::Matrix3d &K, const Sophus::SE3d &T);
+                             const Eigen::Matrix3f &K, const Sophus::SE3d &T);
 
 void VMapToDepth(const cv::cuda::GpuMat vmap, cv::cuda::GpuMat &depth);
